test(i18n): guard against new hardcoded strings in pages and components
